requirements.txt
# Google Cloud & BigQuery
google-cloud-bigquery==3.13.0
google-cloud-bigquery-storage==2.22.0
db-dtypes==1.1.1

# Data Processing
pandas==2.1.3
pyarrow==14.0.1
numpy==1.26.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
flake8==6.1.0
black==23.11.0

# Logging
colorlog==6.8.0
